+++ secondaryFunc.js
@@ -33,3 +33,6 @@ export function delay(interval = 300) {
     });
   }
   
+
+
+
