+++ workWithForm.js
@@ -1,5 +1,9 @@
-import { renderComments, nameInputElement, commentInputElement, mainForm, addComment } from "./script.js";
-import { comments } from "./script.js";
+import { nameInputElement, commentInputElement, mainForm, addComment,pushComment,listElement,comments } from "./script.js";
+
+import { delay, getDate } from "./secondaryFunc.js";
+
+
+
 
 // Функция удаления последнего комментария
 export function delComment() {
@@ -21,8 +25,6 @@ export function delValue() {
   
       if (event.code === "Enter"  ) {
         addComment.click();
-       
-        delValue();
       }
     });  
   }
@@ -38,3 +40,98 @@ export function delValue() {
       });
     });
   }
+
+ // функция Добавление комментария
+export function addcommentuser(){
+    addComment.addEventListener("click", () => {
+  
+      if (nameInputElement.value === "" || commentInputElement.value === "") {
+        
+        nameInputElement.classList.add("error");
+        commentInputElement.classList.add("error");
+        nameInputElement.placeholder = 'Введите имя';
+        commentInputElement.placeholder = 'Введите комментарий';
+        buttonBlock()
+        return;  
+      } 
+    
+        pushComment()
+        nameInputElement.classList.remove("error");
+        commentInputElement.classList.remove("error");
+        const oldListHtml = listElement.innerHTML;
+        renderComments();
+     
+    });
+  }
+
+
+   export function reComment () {
+  
+    const allComments = document.querySelectorAll('.comment')
+    for(let comment of allComments){
+     comment.addEventListener('click', (event)=>{
+       event.stopPropagation()
+       const nameUser = comment.dataset.name;
+       const userComments = comment.dataset.comment;
+       commentInputElement.value =` >${userComments} \n${nameUser} <\n`
+       
+     })
+    
+    }
+   }
+
+// Добавление лайка
+export function addLike () {
+    const likeButtons = listElement.querySelectorAll('.like-button');
+    for(let likeButton of likeButtons){
+  
+      likeButton.addEventListener('click', ( event) => {
+        event.stopPropagation()
+            const index = likeButton.dataset.index;
+            likeButton.classList.add('-loading-like')
+            delay(2000).then(()=> {
+             
+              if (!comments[index].isLiked) {
+                comments[index].isLiked = true;
+                comments[index].likes +=1;
+              } else {
+                comments[index].isLiked = false;
+                comments[index].likes -=1;
+              }
+              renderComments();
+            })
+  
+        })
+    }
+  }   
+
+
+  // Рендер разметки
+
+// Рендер разметки
+export function renderComments() {
+  
+  const userHtml = comments.map((user, index) => {
+    return `<li class="comment"  data-name="${user.author.name}" data-comment="${user.text}">
+    <div class="comment-header">
+      <div>${user.author.name}</div>
+      <div>${getDate(user.date)}</div>
+    </div>
+    <div class="comment-body" >
+   <div class ="comment-text"> ${user.text} </div>
+    </div>
+    <div class="comment-footer">
+      <div class="likes">
+        <span class="likes-counter">${user.likes}</span>
+        <button data-index="${index}"  class="${user.isLiked ? 'like-button -active-like' : 'like-button'}"></button>
+    
+      </div>
+    </div>
+  </li>`
+  }).join('')
+
+  listElement.innerHTML = userHtml;
+  addLike()  // лайки
+  reComment() // Комментирование поста
+
+}
